Wrote new path generator (has performance issue...)
