+++ RobotComponentsABB/Components/Simulation/PathGeneratorComponent.cs
@@ -72,10 +72,9 @@ namespace RobotComponentsABB.Components.Simulation
         private ForwardKinematics _forwardKinematics = new ForwardKinematics();
         private List<Plane> _planes = new List<Plane>();
         private List<Curve> _paths = new List<Curve>();
-        private List<List<double>> _internalAxisValues = new List<List<double>>();
-        private List<List<double>> _externalAxisValues = new List<List<double>>();
+        private List<RobotJointPosition> _robotJointPositions = new List<RobotJointPosition>();
+        private List<ExternalJointPosition> _externalJointPositions = new List<ExternalJointPosition>();
         private int _lastInterpolations = 0;
-        private bool _raiseWarnings = false;
         private bool _previewMesh = true;
         private bool _previewCurve = true;
 
@@ -119,36 +118,27 @@ namespace RobotComponentsABB.Components.Simulation
                 _paths.Clear();
                 _paths = _pathGenerator.Paths;
 
-                // Clear the lists with the internal and external axis values
-                ClearAxisValuesLists();
-                _internalAxisValues = _pathGenerator.InternalAxisValues;
-                _externalAxisValues = _pathGenerator.ExternalAxisValues;
+                // Clear the lists
+                _robotJointPositions.Clear();
+                _externalJointPositions.Clear();
+
+                // Get new lists
+                _robotJointPositions = _pathGenerator.RobotJointPositions;
+                _externalJointPositions = _pathGenerator.ExternalJointPositions;
 
                 // Store the number of interpolations that are used, to check if this value is changed. 
                 _lastInterpolations = interpolations;
-
-                // Raise warnings
-                if (_pathGenerator.ErrorText.Count != 0)
-                {
-                    _raiseWarnings = true;
-                }
-                else
-                {
-                    _raiseWarnings = false;
-                }
             }
 
             // Get the index number of the current target
             int index = (int)(((_planes.Count - 1) * interpolationSlider));
 
-            // Calcualte foward kinematics
-            _forwardKinematics.InternalAxisValues = _internalAxisValues[index];
-            _forwardKinematics.ExternalAxisValues = _externalAxisValues[index];
+            // Calculate foward kinematics
             _forwardKinematics.HideMesh = !_previewMesh;
-            _forwardKinematics.Calculate();
+            _forwardKinematics.Calculate(_robotJointPositions[index], _externalJointPositions[index]);
 
             // Show error messages
-            if (_raiseWarnings == true)
+            if (_pathGenerator.ErrorText.Count != 0)
             {
                 AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, "Only axis values of absolute joint movements are checked.");
 
@@ -168,28 +158,6 @@ namespace RobotComponentsABB.Components.Simulation
             else { DA.SetDataList(4, null); }
         }
 
-        /// <summary>
-        /// This method clears the two lists with internal and external values
-        /// </summary>
-        private void ClearAxisValuesLists()
-        {
-            // Clear the all the individual lists with internal axis values
-            for (int i = 0; i < _internalAxisValues.Count; i++)
-            {
-                _internalAxisValues[i].Clear();
-            }
-
-            // Clear the all the individual lists with external axis values
-            for (int i = 0; i < _externalAxisValues.Count; i++)
-            {
-                _externalAxisValues.Clear();
-            }
-
-            // Clear both primary lists
-            _internalAxisValues.Clear();
-            _externalAxisValues.Clear();
-
-        }
         #region menu item
         /// <summary>
         /// Boolean that indicates if the custom menu item for hiding the robot mesh is checked
